move too default lobby room
