+++ src/views/Entrance/styles.ts
